+++ core/src/stored/dir_cmd.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2001-2012 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2012 Planets Communications B.V.
-   Copyright (C) 2013-2022 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2024 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -219,9 +219,7 @@ static inline bool AreMaxConcurrentJobsExceeded()
   JobControlRecord* jcr;
   unsigned int cnt = 0;
 
-  foreach_jcr (jcr) {
-    cnt++;
-  }
+  foreach_jcr (jcr) { cnt++; }
   endeach_jcr(jcr);
 
   return (cnt >= me->MaxConcurrentJobs) ? true : false;
@@ -424,7 +422,8 @@ static bool SetdebugCmd(JobControlRecord* jcr)
     scan = sscanf(dir->msg, setdebugv0cmd, &level, &trace_flag);
   }
   if ((scan != 3 && scan != 2) || level < 0) {
-    dir->fsend(BADcmd, "setdebug", dir->msg);
+    std::string cpy{dir->msg};
+    dir->fsend(BADcmd, "setdebug", cpy.c_str());
     return false;
   }
 
@@ -462,7 +461,8 @@ static bool SetdeviceCmd(JobControlRecord* jcr)
   int scan = sscanf(dir->msg, setdevice_autoselect, device_name.data(),
                     &autoselect_value);
   if (scan != 2) {
-    dir->fsend(BADcmd, "setdevice", dir->msg);
+    std::string cpy{dir->msg};
+    dir->fsend(BADcmd, "setdevice", cpy.c_str());
     return false;
   }
 
@@ -556,10 +556,8 @@ static bool CancelCmd(JobControlRecord* cjcr)
     ReleaseDeviceCond();
   }
 
-  /*
-   * See if the Job has a certain protocol.
-   * When canceling a NDMP job make sure we call the end_of_ndmp_* functions.
-   */
+  /* See if the Job has a certain protocol.
+   * When canceling a NDMP job make sure we call the end_of_ndmp_* functions. */
   switch (jcr->getJobProtocol()) {
     case PT_NDMP_BAREOS:
       switch (jcr->getJobType()) {
@@ -630,14 +628,12 @@ static bool DoLabel(JobControlRecord* jcr, bool relabel)
   bool ok = false;
   slot_number_t slot;
 
-  /*
-   * Determine the length of the temporary buffers.
+  /* Determine the length of the temporary buffers.
    * If the total length of the incoming message is less
    * then MAX_NAME_LENGTH we can use that as the upper limit.
    * If the incomming message is bigger then MAX_NAME_LENGTH
    * limit the temporary buffer to MAX_NAME_LENGTH bytes as
-   * we use a sscanf %127s for reading the temorary buffer.
-   */
+   * we use a sscanf %127s for reading the temorary buffer. */
   len = dir->message_length + 1;
   if (len > MAX_NAME_LENGTH) { len = MAX_NAME_LENGTH; }
 
@@ -1360,10 +1356,8 @@ static bool ChangerCmd(JobControlRecord* jcr)
   const char* cmd = NULL;
   bool ok = false;
   bool is_transfer = false;
-  /*
-   * A safe_cmd may call autochanger script but does not load/unload
-   *    slots so it can be done at the same time that the drive is open.
-   */
+  /* A safe_cmd may call autochanger script but does not load/unload
+   *    slots so it can be done at the same time that the drive is open. */
   bool safe_cmd = false;
 
   if (sscanf(dir->msg, "autochanger listall %127s", devname.c_str()) == 1) {
@@ -1617,7 +1611,8 @@ class ReplicateCmdConnectState {
         || state_ == ReplicateCmdState::kError) {
       if (!jcr_) { return; }
       if (jcr_->dir_bsock) {
-        jcr_->dir_bsock->fsend(BADcmd, "replicate", jcr_->dir_bsock->msg);
+        std::string cpy{jcr_->dir_bsock->msg};
+        jcr_->dir_bsock->fsend(BADcmd, "replicate", cpy.c_str());
       }
       jcr_->store_bsock = nullptr;
     }
@@ -1643,7 +1638,8 @@ static bool ReplicateCmd(JobControlRecord* jcr)
   if (sscanf(dir->msg, replicatecmd, &JobId, JobName, stored_addr, &stored_port,
              &tls_policy, sd_auth_key.c_str())
       != 6) {
-    dir->fsend(BADcmd, "replicate", dir->msg);
+    std::string cpy{dir->msg};
+    dir->fsend(BADcmd, "replicate", cpy.c_str());
     return false;
   }
 
@@ -1738,11 +1734,13 @@ static bool PassiveCmd(JobControlRecord* jcr)
   char filed_addr[MAX_NAME_LENGTH];
   BareosSocket* dir = jcr->dir_bsock;
   BareosSocket* fd; /* file daemon bsock */
+  std::string cpy{dir->msg};
 
-  Dmsg1(100, "PassiveClientCmd: %s", dir->msg);
-  if (sscanf(dir->msg, passiveclientcmd, filed_addr, &filed_port, &tls_policy)
+  Dmsg1(100, "PassiveClientCmd: %s", cpy.c_str());
+  if (sscanf(cpy.c_str(), passiveclientcmd, filed_addr, &filed_port,
+             &tls_policy)
       != 3) {
-    PmStrcpy(jcr->errmsg, dir->msg);
+    PmStrcpy(jcr->errmsg, cpy.c_str());
     Jmsg(jcr, M_FATAL, 0, _("Bad passiveclientcmd command: %s"), jcr->errmsg);
     goto bail_out;
   }
@@ -1807,7 +1805,7 @@ static bool PassiveCmd(JobControlRecord* jcr)
   return dir->fsend(OKpassive);
 
 bail_out:
-  dir->fsend(BADcmd, "passive client");
+  dir->fsend(BADcmd, "passive client", cpy.c_str());
   return false;
 }
 
@@ -1815,10 +1813,11 @@ static bool PluginoptionsCmd(JobControlRecord* jcr)
 {
   BareosSocket* dir = jcr->dir_bsock;
   char plugin_options[2048];
+  std::string cpy{dir->msg};
 
-  Dmsg1(100, "PluginOptionsCmd: %s", dir->msg);
-  if (sscanf(dir->msg, pluginoptionscmd, plugin_options) != 1) {
-    PmStrcpy(jcr->errmsg, dir->msg);
+  Dmsg1(100, "PluginOptionsCmd: %s", cpy.c_str());
+  if (sscanf(cpy.c_str(), pluginoptionscmd, plugin_options) != 1) {
+    PmStrcpy(jcr->errmsg, cpy.c_str());
     Jmsg(jcr, M_FATAL, 0, _("Bad pluginoptionscmd command: %s"), jcr->errmsg);
     goto bail_out;
   }
@@ -1833,7 +1832,7 @@ static bool PluginoptionsCmd(JobControlRecord* jcr)
   return dir->fsend(OKpluginoptions);
 
 bail_out:
-  dir->fsend(BADcmd, "plugin options");
+  dir->fsend(BADcmd, "plugin options", cpy.c_str());
   return false;
 }